planete et equipement ok, pouvoir et vehicule fait mais a tester. Fonction pour chemin image ok

+++ src/utils/suppressionUtils.js
@@ -13,17 +13,26 @@ async function supprimerImagesEtAttributsParEntite(entiteType, entiteId) {
             }
         });
 
-        // Supprimer les images associées
+        // Récupérer toutes les images associées à cette entité
         const images = await Image.findAll({
             where: { entiteType, entiteId }
         });
 
+        if (images.length === 0) {
+            console.log(`Aucune image trouvée pour ${entiteType} avec l'ID ${entiteId}.`);
+            return;
+        }
+
+        // Si des images sont trouvées, les supprimer
         for (const image of images) {
-            const imagePath = path.join(__dirname, '..', image.chemin); // Suppression du fichier image
-            if (fs.existsSync(imagePath)) {
-                fs.unlinkSync(imagePath);
+            if (image && image.chemin) {
+                const imagePath = path.join(__dirname, '..', image.chemin); // Suppression du fichier image
+                if (fs.existsSync(imagePath)) {
+                    fs.unlinkSync(imagePath);
+                    console.log(`Image supprimée du serveur: ${imagePath}`);
+                }
+                await image.destroy(); // Supprimer l'image en base de données
             }
-            await image.destroy(); // Supprimer l'image en base de données
         }
     } catch (error) {
         console.error(`Erreur suppression images et attributs pour ${entiteType} ${entiteId}:`, error);
